+++ metar_taf_parser/command/metar.py
@@ -1,8 +1,9 @@
 import re
 
 from metar_taf_parser.commons import converter
-from metar_taf_parser.model.enum import DepositType, DepositThickness, DepositCoverage, DepositBrakingCapacity
+from metar_taf_parser.model.enum import DepositType, DepositCoverage
 from metar_taf_parser.model.model import RunwayInfo, Metar
+from metar_taf_parser.commons.i18n import _
 
 
 class AltimeterCommand:
@@ -51,6 +52,26 @@ class RunwayCommand:
         self._max_range_pattern = re.compile(RunwayCommand.runway_max_range_regex)
         self._runway_pattern = re.compile(RunwayCommand.runway_regex)
         self._runway_deposit_pattern = re.compile(RunwayCommand.runway_deposit_regex)
+        self._deposit_thickness = {
+            '//': 'DepositThickness.//',
+            '00': 'DepositThickness.00',
+            '92': 'DepositThickness.92',
+            '93': 'DepositThickness.93',
+            '94': 'DepositThickness.94',
+            '96': 'DepositThickness.96',
+            '97': 'DepositThickness.97',
+            '98': 'DepositThickness.98',
+            '99': 'DepositThickness.99'
+        }
+        self._deposit_braking_capacity = {
+            '//': 'DepositBrakingCapacity.//',
+            '91': 'DepositBrakingCapacity.91',
+            '92': 'DepositBrakingCapacity.92',
+            '93': 'DepositBrakingCapacity.93',
+            '94': 'DepositBrakingCapacity.94',
+            '95': 'DepositBrakingCapacity.95',
+            '99': 'DepositBrakingCapacity.99'
+        }
 
     def can_parse(self, input: str):
         return self._generic_pattern.match(input)
@@ -62,8 +83,8 @@ class RunwayCommand:
             runway.name = matches[0][0]
             runway.deposit_type = DepositType(matches[0][1])
             runway.coverage = DepositCoverage(matches[0][2])
-            runway.thickness = DepositThickness(matches[0][3])
-            runway.braking_capacity = DepositBrakingCapacity(matches[0][4])
+            runway.thickness = self.__parse_deposit_thickness(matches[0][3])
+            runway.braking_capacity = self.__parse_deposit_braking_capacity(matches[0][4])
             metar.add_runway_info(runway)
             return
 
@@ -83,6 +104,14 @@ class RunwayCommand:
             runway.max_range = int(matches[0][2])
             runway.trend = matches[0][3]
             metar.add_runway_info(runway)
+
+    def __parse_deposit_thickness(self, input):
+        thickness = self._deposit_thickness.get(input, 'DepositThickness.default')
+        return _(thickness).format(input)
+
+    def __parse_deposit_braking_capacity(self, input):
+        braking_capacity = self._deposit_braking_capacity.get(input, 'DepositBrakingCapacity.default')
+        return _(braking_capacity).format(float(input) / 100)
 
 
 class TemperatureCommand: